Buttons components modified
Changed API for StepContainer with Back and Next buttons on <StepWithNextAndBackContainer> component. An object for configurating each button's props is now expected. If none is provided, the default behaviour will be set.

NextButton and BackButton components changed to meet the Container's changes.

A new type is now exported from the button's root file.

TODO: Ideally, a single Button component should be created, that will export a basic button's schema type, which will then be used to create either a Primary or a Secondary button. This is preferable over having two different components that only may differ in stylings.

// Maybe a good opportunity to incorporate StyledComponents library???